Make IDs 32bits (#542)
Step towards #141

When thinking about using u32 IDs, I was getting really caught up by the
fact that Rust's `HashMap` implementation still requires storing a u64
for the key digests. Of course that consumes more memory and ideally we
can have a custom `HashMap` implementation that uses half the amount of
bits.

However, we actually use IDs quite extensively outside of just `HashMap`
keys. For example, many definitions store StringId, names store multiple
NameIds and ancestors is a list of DeclarationId. We can actually have
very substantial memory savings right now even without the 32 bit
`HashMap`.

### Implementation

This PR reduces the size of our IDs to u32 from i64. I recommend
reviewing per commit:

1. Reduce the ID size to u32
2. Allow `DefinitionId` and `ReferenceId` to be tagged with their kind.
This helps us reduce the changes of collision by encoding the kind in
the lower bits of the ID. Essentially, we have 28 bits for the digest +
4 for the kind. Definitions and references are the things we have the
most in Core and the only ones I got conflicts for running the release
mode, so I think this is enough for the time being.

As an added benefit, this allows to check the kind of definitions and
references without having to retrieve them from the graph. That
information is directly encoded in the ID, we can just invoke `kind`.
Note that the `kind` addition conflicts with #502, we can let that PR go
first.
3. Adjust the C side to use u32

### Impact

Despite still using 64 bit `HashMaps`, this PR still reduces our memory
used from `~3810 MB` to
`~2880 MB` (25% reduction).

I believe we can further reduce this with a custom `HashMap`
implementation that stores u32 internally.

### Why now?

I believe this is the right time to do this for the following reason:
we're starting adoption of Rubydex in our tools. It will be easier for
us to verify that we can indeed get away with u32 and lower memory usage
_and then_ increase to u64 if necessary than the other way around.

Also, we're almost getting to 4GB and a reduction is definitely welcome.

+++ rust/rubydex-sys/src/declaration_api.rs
@@ -21,7 +21,7 @@ use rubydex::model::ids::{DeclarationId, StringId};
 ///
 /// This function will panic if the name pointer is invalid.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_declaration_name(pointer: GraphPointer, name_id: i64) -> *const c_char {
+pub unsafe extern "C" fn rdx_declaration_name(pointer: GraphPointer, name_id: u32) -> *const c_char {
     with_graph(pointer, |graph| {
         let name_id = DeclarationId::new(name_id);
         if let Some(decl) = graph.declarations().get(&name_id) {
@@ -40,9 +40,9 @@ pub unsafe extern "C" fn rdx_declaration_name(pointer: GraphPointer, name_id: i6
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn rdx_declaration_member(
     pointer: GraphPointer,
-    name_id: i64,
+    name_id: u32,
     member: *const c_char,
-) -> *const i64 {
+) -> *const u32 {
     let Ok(member_str) = (unsafe { utils::convert_char_ptr_to_string(member) }) else {
         return ptr::null();
     };
@@ -72,7 +72,7 @@ pub unsafe extern "C" fn rdx_declaration_member(
 ///
 /// This function will panic if the name pointer is invalid.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn rdx_declaration_unqualified_name(pointer: GraphPointer, name_id: i64) -> *const c_char {
+pub unsafe extern "C" fn rdx_declaration_unqualified_name(pointer: GraphPointer, name_id: u32) -> *const c_char {
     with_graph(pointer, |graph| {
         let name_id = DeclarationId::new(name_id);
         if let Some(decl) = graph.declarations().get(&name_id) {
@@ -96,7 +96,7 @@ pub unsafe extern "C" fn rdx_declaration_unqualified_name(pointer: GraphPointer,
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn rdx_declaration_definitions_iter_new(
     pointer: GraphPointer,
-    decl_id: i64,
+    decl_id: u32,
 ) -> *mut DefinitionsIter {
     // Snapshot the IDs and kinds at iterator creation to avoid borrowing across FFI calls
     with_graph(pointer, |graph| {
@@ -104,7 +104,7 @@ pub unsafe extern "C" fn rdx_declaration_definitions_iter_new(
         if let Some(decl) = graph.declarations().get(&decl_id) {
             rdx_definitions_iter_new_from_ids(graph, decl.definitions())
         } else {
-            DefinitionsIter::new(Vec::<(i64, DefinitionKind)>::new().into_boxed_slice())
+            DefinitionsIter::new(Vec::<(u32, DefinitionKind)>::new().into_boxed_slice())
         }
     })
 }
